requirements.txt
Flask==2.2.5
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.5
Flask-Login>=0.6.3
Flask-WTF>=1.2.0
Flask-Cors>=4.0.0
Flask-Caching>=2.1.0
Flask-Limiter>=3.0.0
flask-restx==1.3.0
sentry-sdk[flask]>=1.30.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
alembic>=1.12.0
Werkzeug>=3.0.0
Jinja2>=3.1.3
WTForms>=3.1.0
pytest>=8.0.0
coverage[toml]>=7.5.0
Flask-Babel==3.1.0
